+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.16.0 (TBD)
 
 - [BREAKING] Incremented MSRV to 1.88.
+- [BREAKING] Remove generics from `MerkleStore` and remove `KvMap` and `RecordingMap` (#442).
+- Rename `smt_hashmaps` feature to `hashmaps` (#442).
 
 ## 0.15.3 (2025-06-18)